% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome}
\alias{chromosome}
\title{Human chromosome summary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 16 columns.}
\source{
\url{https://en.wikipedia.org/wiki/Human_genome}
}
\usage{
data(chromosome)
}
\description{
A dataset containing summary information of the human chromosomes
}
\keyword{datasets}
