% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61187 rows and 10 columns.}
\source{
\url{https://www.ensembl.org/index.html}
}
\usage{
data(genes)
}
\description{
A dataset containing Ensembl's unique genes in the Human Genome (GRCh38.p13) and other associated information.
}
\keyword{datasets}
