% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id}
\alias{id}
\title{id}
\format{
A data frame with 26,445 rows and 5 variables:
\describe{
  \item{hgnc_id}{A unique ID provided by the HGNC for each gene with an approved symbol. IDs are of the format HGNC:n, where n is a unique number. HGNC IDs remain stable even if a name or symbol changes}
  \item{approved_symbol}{The official gene symbol approved by the HGNC, which is typically a short form of the gene name. Symbols are approved in accordance with the Guidelines for Human Gene Nomenclature}
  \item{approved_name}{The full gene name approved by the HGNC; corresponds to the approved symbol above}
  \item{aka}{This is a symbol or name by which a gene has been alternatively referred to in the literature or databases, or which groups it into a known gene family. Synonyms are recorded along with the approved symbols and names as part of the gene entry to facilitate database searching}
  \item{chromosome}{Indicates the cytogenetic location of the gene or region on the chromosome}
  \item{ref_seq_i_ds}{The Reference Sequence (RefSeq) identifier for that entry, provided by the NCBI. As we do not aim to curate all variants of a gene only one selected RefSeq is displayed per gene report. RefSeq aims to provide a comprehensive, integrated, non-redundant set of sequences, including genomic DNA, transcript (RNA), and protein products}
  \item{ncbi_gene_id}{The NCBI Gene ID refers to page at the NCBI provides curated sequence and descriptive information about genetic loci including official nomenclature, synonyms, sequence accessions, phenotypes, EC numbers, MIM numbers, UniGene clusters, homology, map locations, and related web sites}
  \item{locus_type}{Specifies the genetic class of each gene entry}
  \item{omim_id_supplied_by_omim}{ID links to the Online Mendelian Inheritance in Man page for the gene. OMIM is described as a catalog of human genes and genetic phenotypes containing textual information, and links to MEDLINE, sequence records in the Entrez system, and additional related resource}
  \item{uni_prot_id_supplied_by_uni_prot}{The UniProt page for the encoded gene protein product. The UniProt Protein Knowledgebase is a curated protein sequence database that provides a high level of annotation, a minimal level of redundancy and high level of integration with other databases}
  \item{entrez_summary}{NCBI Entrez Gene summary manually curated and provided by RefSeq}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov}
}
\usage{
data(id)
}
\description{
A dataset with gene summary information, as well as linking MIM numbers, NCBI Gene IDs, Uniprot IDs, and HGNC Approved Gene Symbols.
}
\keyword{datasets}
