% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{join}
\alias{join}
\title{join}
\format{A data frame with 26,445 rows and 5 variables:
\describe{
  \item{gene_name}{Official gene symbol from HGNC.}
  \item{gene_id_entrez}{Official Entrez gene ID from NCBI.}
  \item{gene_id_ensembl}{Official gene ID from Ensembl.}
  \item{mim_number}{Each OMIM entry is given a unique six-digit number.}
  \item{mim_entry_type}{Gene indicates that the entry contains the description of a gene of known sequence and a phenotype. See other types at faq 1.3 at https://omim.org help faq for more types.}
}}
\source{
\url{https://www.omim.org/downloads/}
}
\usage{
data(join)
}
\description{
A dataset linking MIM numbers, NCBI Gene IDs, Ensembl Gene IDs, and HGNC Approved Gene Symbols.
}
\keyword{datasets}
