% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_class.R
\name{launch_class}
\alias{launch_class}
\title{Launch Rmd files for in-class activities}
\usage{
launch_class(class = 1)
}
\arguments{
\item{class}{enter the relevant class number (as an int)}
}
\value{
opens the corresponding Rmd file in RStudio
}
\description{
This function launches the named class Rmd file
}
\author{
Akshay Bareja
}
