% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_homework.R
\name{launch_homework}
\alias{launch_homework}
\title{Launch Rmd files for homework}
\usage{
launch_homework(class = 1)
}
\arguments{
\item{class}{enter the relevant class number (as an int)}
}
\value{
opens the corresponding Rmd file in RStudio
}
\description{
This function launches the named class Rmd file
}
\author{
Akshay Bareja
}
