% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omim_genemap}
\alias{omim_genemap}
\title{omim_genemap}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16992 rows and 14 columns.}
\source{
\url{https://www.omim.org/downloads/}
}
\usage{
data(omim_genemap)
}
\description{
A dataset containing OMIM's Synopsis of the Human Gene Map, including additional information such as genomic coordinates and inheritance.
}
\keyword{datasets}
