% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omim_mim2gene}
\alias{omim_mim2gene}
\title{omim_mim2gene}
\format{A data frame with 26,445 rows and 5 variables:
\describe{
  \item{number_mim_number}{Each OMIM entry is given a unique six-digit number.}
  \item{mim_entry_type}{Gene indicates that the entry contains the description of a gene of known sequence and a phenotype. See other types at faq 1.3 at https://omim.org help faq for more types.}
  \item{entrez_gene_id_ncbi}{Official Entrez gene ID from NCBI.}
  \item{approved_gene_symbol_hgnc}{Official gene symbol from HGNC.}
  \item{ensembl_gene_id_ensembl}{Official gene ID from Ensembl.}
}}
\source{
\url{https://www.omim.org/downloads/}
}
\usage{
data(omim_mim2gene)
}
\description{
A dataset linking MIM numbers with NCBI Gene IDs, Ensembl Gene IDs, and HGNC Approved Gene Symbols. For more informaation, see: https://omim.org/help/faq.
}
\keyword{datasets}
