% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omim_mimtitles}
\alias{omim_mimtitles}
\title{omim_mimtitles}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26350 rows and 5 columns.}
\source{
\url{https://www.omim.org/downloads/}
}
\usage{
data(omim_mimtitles)
}
\description{
A dataset of MIM numbers and perferred titles.
}
\keyword{datasets}
