% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omim_morbidmap}
\alias{omim_morbidmap}
\title{omim_morbidmap}
\format{A data frame with 7,831 rows and 4 variables:
\describe{
  \item{number_phenotype}{Each Phenotype is followed by its MIM number, if different from that of the locus/gene, and then followed by its phenotype mapping key in parentheses.}
  \item{gene_symbols}{Gene symbol(s) associated with disease}
  \item{mim_number}{Each OMIM entry is given a unique six-digit number.}
  \item{cyto_location}{Computed cyto location from UCSC}
  }}
\source{
\url{https://www.omim.org/downloads/}
}
\usage{
data(omim_morbidmap)
}
\description{
A dataset containing OMIM's Synopsis of the Human Gene Map (same as omim_genemap) sorted alphabetically by disorder.
}
\keyword{datasets}
