% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteins}
\alias{proteins}
\title{proteins}
\format{A data frame with 20,430 rows and 8 variables:
\describe{
  \item{uniprot_id}{Official uniprot entry id}
  \item{gene_name}{Gene name}
  \item{gene_name_alt}{Alternative gene names associated with entry}
  \item{protein_name}{Protein name}
  \item{protein_name_alt}{Alternative protein names associated with entry}
  \item{sequence}{Primary amino acid sequencxe}
  \item{length}{Number of amino acids}
  \item{mass}{Molecular weight of protein}
  }}
\usage{
data(proteins)
}
\description{
A dataset containing human proteome information from Uniprot.
}
\keyword{datasets}
