% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_provenance.R
\name{session_provenance}
\alias{session_provenance}
\title{A Session Provenance Function}
\usage{
session_provenance()
}
\description{
This function allows you to easilty print relevant session information, but not unloaded packages, for provenance and reproducibility.
}
\author{
Matthew Hirschey, \email{matthew@hirschey.org}
}
\keyword{sessionInfo}
