% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybiology.R
\docType{package}
\name{tidybiology}
\alias{tidybiology}
\alias{tidybiology-package}
\title{tidybiology: An Introduction to Biological Data Science in R}
\description{
Supporting materials for a course on data science. It contains several sample data sets and some utility functions for data analysis and graphing.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tidybiology.org}
}

}
\author{
Matthew Hirschey \email{matthew@hirschey.org}
}
