% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subcell}
\alias{subcell}
\title{subcell}
\format{A data frame with 20,840 rows and 5 variables:
\describe{
  \item{ensembl_prot_id}{Official Ensembl protein ID}
  \item{gene_name}{Official gene name}
  \item{go_term}{Gene Ontology Term}
  \item{score}{Integrated subcellular location score}
  \item{location}{Subcellular Location}
  }}
\source{
\url{https://compartments.jensenlab.org/Downloads}
}
\usage{
data(subcell)
}
\description{
A dataset containing cleaned protein subcellular localization information.
}
\keyword{datasets}
